fix(build): Resolve integration test compilation failure
The integration test build was failing due to a Go build tag issue that caused the `isPublicURL` function to be declared twice.

The initial attempt to fix this by adding `//go:build !integration` to the top of `pkg/converter/converter.go` was incorrect. This excluded the entire file, including the `Converter` struct definition, which led to an `undefined: Converter` error in `ssrf_mock.go`.

This commit resolves the issue by isolating the conditionally compiled code:

- The production `isPublicURL` function has been moved from `converter.go` to a new `pkg/converter/ssrf.go` file. This new file is marked with the `//go:build !integration` tag.
- The mock `isPublicURL` function remains in `pkg/converter/ssrf_mock.go`, which is correctly marked with `//go:build integration`.

This ensures that the main `Converter` struct is always included in the build, while the appropriate implementation of `isPublicURL` is selected based on the presence of the `integration` build tag.
